Add an initial, partial prototype broker implementation.
The broker.py file is the broker, it supports only creating and joining
rooms, along with status information.

The ws-room.py script creates a new room on the broker, and then shows
status updates.

The ws-join.py script joins a room and then shows status updates.

Both disconnect when ^C is pressed.